Update environment variable name
We have agreed to use more descriptive environment variable names
for clarity across the different applications.
All are prefixed with 'DM_'
Next the name of the application they apply to - 'SEARCH_API' in this case.
Next a description of it's purpose - i.e. 'AUTH_TOKENS'

So in this case we have DM_SEARCH_API_AUTH_TOKENS instead of simply AUTH_TOKENS as it was previously.