Fix type error caused by error reporting
Currently some error handling code tries to get `exception.info["error"]` but
for some types of exception `exception.info` is a list, not a map, so this results in:
`TypeError: sequence index must be integer, not 'str'`
when reporting errors, and the original error message is lost.

This adds a method that should ensure that there is some message to include for error
reporting, without causing a further error to be thrown.