Replace query_string with simple_query_string
[#96078646](https://www.pivotaltracker.com/story/show/96078646)

Fixes error responses caused by search requests with the invalid
lucene query syntax by using the simplified query type that only
supports the basic operators and discards invalid parts of the
query instead of throwing an exception.

simple_query_string is configured to support AND, OR, NOT and quote
operators.

Unlike query_string it doesn't seem to support the dis_max query,
which might change the result order. It's hard to say how much this
will affect the results, but using AND as default operator narrows
the results with each added term, so the ranking is less noticeable
than it would be with "OR" queries. It doesn't look like the Grails
app is using the dis_max either.