Add datacentreTier to the list of filter fields
Datacentre Tier is one of the filters shown in the buyer-frontend,
so it should be indexed and accepted as a filter field.

Right now adding a field requires adding it to the list of fields
and to the mapping attached to the index. The existing index mapping
is not affected by the mapping.json change.