Disallow implicit field creation in the services mapping
By default elasticsearch will try to guess the types of fields that
aren't listed in the index mapping and will add them to the mapping.

Setting dynamic to strict makes ES insted throw an error when trying
to index a document with unknown fields. This makes sure that our
mapping is up-to-date and the indexed documents are matching it.