Add rough word limit on es highlight excerpts
Highlight excerpts for all service fields returned from elasticsearch
now 400 character limit. This is enforced whether or not a search
term has been found.
If no search term is found, up to the first 400 characters of the field
will be used as the basis for the `highlight` snippet.
elasticsearch tries not to break in the middle of a word.  From the
documentation: "The actual length may be shorter than specified as it tries
to break on a word boundary."
Trailing punctuation is also removed.
Roughly approximates our 50-word limit on service summaries for G6+ services.

More information about highlighted fragments can be found [here](https://www.elastic.co/guide/en/elasticsearch/reference/1.6/search-request-highlighting.html#_highlighted_fragments)

Discovered what appears to be a bug.
The character limit for a highlight excerpt (when a match has been found) is
only respected if the character limit is less than ~2/3 of the original data.
For example, if a matched word is found in a 600-character service summary
and our `fragment_size` limit is set to `400`, the entire summary will be
returned as a highlight fragment.
A 610-character service summary would be truncated as expected.
Using the fast vector highlighter seems to fix the problem, but since it's
not really an issue for us, I've not changed the default highlighter.