Use postings highlighter for the serviceSummary field
[#103575496](https://www.pivotaltracker.com/story/show/103575496)

Changes the serviceSummary mapping to make Elasticsearch use `postings`
highlighter for the service description field. Postings highlighter
fixes bugs with the unhighlighted document HTML escaping and fragment
size limits.

(https://www.elastic.co/guide/en/elasticsearch/reference/2.0/search-request-highlighting.html#postings-highlighter)

The default "plain" highlighter ignores the encoder option when nothing
was highlighted in the output, returning the truncated field text with
unescaped HTML entities, which creates an XSS vulnerability in search
results.

We could ignore the highlighter result and instead escape the original
description ourselves in this case, but we also use the highlighter to
limit the search result snippet size.

Instead, we can set a different highlighter by specifying a field option
in the index mapping.

Both `postings` and `fvh` highlighters properly encode unhighlighted
snippets. They also fix a previous issue with fragment size limit we've
run into in #51.

Setting a different highlighter requires a change to the mapping.
For `fvh` this change results in a failure to update existing index
in-place. For `postings` Elasticsearch acknowledges the mapping update,
but doesn't actually change the field option and doesn't switch the type
of the highlighter.

So this mapping change won't be applied until the index is recreated.