Fix search query tests by explicitly creating the search index mapping
Not creating an index before inserting documents created a dynamic
mapping for the test index. Creating the index explicitly sets the
services mapping and seems to change the simple query behaviour with
some special characters (&, && and |). There doesn't seem to be a flag
setting to change this behaviour and it matches the current search
results in production. I've updated the test to document the current
behaviour for now. We can add additional logic to work around the
special characters and update the tests if needed later.