Install libffi-dev on the ElasticBeanstalk environments
We've added cffi to requirements.txt for PaaS, since dmutils depends
on it.

PaaS requires specifying python dependencies that require a C library
in the application requirements.txt, which it parses to install the
required system packages.

For Elastic Beanstalk we need to specify the list of packages
manually in .ebextensions