Add `update_conditionally` transformation
A transformation to conditionally update the value of a field.

This is being created so that the value of the status field for a brief
being indexed can be changed from cancelled, withdrawn or awarded to
just closed.

We will be rolling out the new statuses, but for now we just want to
maintain existing functionality.