Handle and log string status codes
In api_response we were catching TypeErrors when checking the status
code returned from elasticsearch. Elasticsearch has a weird thing where
it can return 'N/A' as the status code.

Previously we were checking if the status code was 'N/A' and handling it
if it was. If it wasn't we just reraised the TypeError.

This meant that we lost any info about the error returned by
Elasticsearch, and only got the stack trace leading to the TypeErrorin
the app.

We had a bunch of 500's caused by this, as Elasticsearch returned
something with a string as it's status code that wasn't 'N/A'. Due to
our lack of logging we don't know what it was.

This update will allow us to handle any string status codes, and log
them so we know what they are.