remove ability to update index mapping through attempting to "create" an existing index
The existing implementation had the flaw that it failed to update the
`settings` accompanying an updated mappings file. Fixing that proves to
be quite tricky, mostly because ES will refuse to accept the attempted
setting of "static" `settings` on a live index, even if they are unchanged
from their previous values. This means we must determine which settings
don't match those already on the index, as attempting to put any of them
that are considered "static" would needlessly cause the update to fail. The
difficulty in doing *that* comes from ES's habit of allowing deep-nested
json object references `to.be.specified.like.this`, and in places allowing
omission of levels of the nesting entirely. These things would all have to
be normalized out and it seems like more code weight than it's worth, given
that the feature still really doesn't buy us much - the few things that
you're able to update on a live index generally require you to reindex
anyway to gain the benefit from them.