core_search_and_aggregate: catch situations where a 500 is returned because max_result_window is exceeded
in most cases these should be treated as a 404 but only after we've checked
that we are actually requesting a page that wouldn't exist given the result
count or whether we just haven't set the `max_result_window` high enough
for this index. it does mean firing off another request just to be able to
categorize the error, but that's probably worth it to be able to return
something sane (and leave something less alarming in our logs).