Add field for filtering open and closed briefs
We would like to show more detail about the status of an award in the
search results for briefs. Currently the statuses are transformed in the
search mapping to either 'open' or 'closed'.

We need to maintain the original status values for each brief in
Elasticsearch whilst also retaining the current filtering ability.

This PR adds a brand new target field `statusOpenClosed` which will be
used by the `buyer-frontend` to continue to filter on 'open' and
'closed' briefs.

We are keeping the old transformations of the status field in place for
the moment until the updates to the `buyer-frontend` have been made.