Find elasticsearch URI in VCAP_SERVICES by tag
GOV.UK PaaS are planning to migrate from using Docker compose with 
elasticsearch. In preparation for this they changed the VCAP_SERVICE 
label from 'elasticsearch' to 'elasticsearch-compose' see (#153). In 
future, they will be changing the label back from 
'elasticsearch-compose' to 'elasticsearch'.

This commit changes the way that we find the elasticsearch entry in 
VCAP_SERVICES to be robust to these changes; assuming that the service 
always contains the tag 'elasticsearch', 
`get_service_by_tag_from_vcap_services` should be able to find it.