Bump utils to 40.9.1
It pulls in a temporary (hopefully) fix to not configure the
urllib3.util.retry logger for the search api.

The Elasticsearch client causes it to generate a lot of unnecessary
logs. We're going to put a PR in upstream and see if we can get it fixed
there, but we'll go with this for now.