Find elasticsearch URI in VCAP_SERVICES by name
GOV.UK PaaS are planning to migrate from using Docker compose with 
elasticsearch. In preparation for this they changed the VCAP_SERVICE 
label from 'elasticsearch' to 'elasticsearch-compose' see (#153). In 
future, they will be changing the label back from 
'elasticsearch-compose' to 'elasticsearch'.

This commit changes the way that we find the elasticsearch entry in 
VCAP_SERVICES to be robust to these changes; the service will (according 
to PaaS) always contains the name of the service as defined in the 
CloudFoundary console ('search_api_elasticsearch' in this case), so 
`get_service_by_name_from_vcap_services` should be able to find it.