Add a switch block to aid in migrating ElasticSearch provider
Change the application init so it takes different arguments from VCAP
services based on the service provider.

We detect the service provider based on the tags part of the service
object.

This code can be removed once migration is fully completed.

This is part of the [ElasticSearch migration ticket](https://trello.com/c/k6OgT7v4)