Add NotFoundError exception handler for get_mapping
At some point between 5.4 and 5.6 (probably [this commit][1])
Elasticsearch stopped returning an empty response for a bad doc type and
started returning a 404.

This commit fixes the get_mapping function so it correctly detects the
case where a 404 has been raised due to a bad doc type, put still passes
the exception up the chain if not so it can be handled by the
`core_search_and_aggregate` function.

[1]: https://github.com/elastic/elasticsearch/commit/dc38e666628512b63c06a78a73531fc5f260c434