Refactor tests of highlighting
We rely heavily on the Elasticsearch highlighting engine in the buyer
frontend, as we show the highlighted text in search results always
(regardless of whether any search terms were matched) and we use its
HTML escaping method as a form of protection against XSS.

However, our tests of the highlighted response have not kept up-to-date,
so this commit shows them a little love.

Tests of the highlighted service description have been moved from
`TestSearchEndpoint` to their own class, `TestHightlightedService`, and
significantly rewritten and expanded.

The aim of this was to make the tests more clearly reflect our
requirements of the highlight field, and make sure that any changes
brought in by Elasticsearch 6 are caught.