Add workaround for Elasticsearch issue #41066
This is a quick hack to make the highlighting work the way we want with
Elasticsearch server >6.0.0,<6.7.2. Hopefully it can be removed soon.

The workaround pushes the complexity of `core_search_and_aggregate`
above the limit, so a noqa directive has been added to tell flake8 to
give this function a pass for now; it should be removed with the
workaround.