Allow overriding development server port
There are times when we might want to change the port that the search
api server listens on in development mode; this commit makes this easier
by adding an additional configuration variable `DM_SEARCH_API_PORT` that
can be set in the environment.

This commit adds `DM_SEARCH_API_PORT` to the `.flaskenv` configuration
file; this works because `python-dotenv` does not override existing
environment variables.

We then use `DM_SEARCH_API_PORT` as the port for the development server
by setting `FLASK_RUN_PORT`; see [the Flask cli docs][1] for more
details on how the development server is configured.

We have split the configuration out in this way because the intention
is that other apps that expect to communicate with the search api will
also look at `DM_SEARCH_API_PORT` when configuring themselves in
development mode; this way we can change the entire system by changing
one environment variable.

[1]: http://flask.palletsprojects.com/en/1.0.x/cli/