Add document type names to index _meta
Although we don't use multiple document types, our API still relies on
document types in the design of its URLs. We can maintain backwards
compatibility by keeping those document types around; so the the API
knows what the document type is supposed to be we add a key to the
`_meta` object in the mapping defintition.