Revert "Add a test to indicate/ ensure that flask is performing retries  …"
This reverts commit 9dab9c08c57ab0548beaa32765a2f064d2ec6544.

This test no longer works under Elasticsearch 7.14.0. This is because Elasticsearch now makes a single request to '/' as part of its setup. It uses this to check it is talking to a valid Elasticsearch, but doesn't retry on failure.

I propose to remove this test because I don't think it's a test that we should have. This test tests Elasticsearch - not any of our code - and reaches deep into the implementation details to mock out the response. This means that the test is very brittle to changes in the Elasticsearch implementation. It also delivers no value, since it is already the responsability of Elasticsearch to test its own features.